HD_Biomarkers

HABS-HD dataset.

Usage

data(HD_Biomarkers)

Format

A data frame with 4712 rows and 73 variables.

Details

Visit_ID: Visit number. Units: -9999. Coding: Numeric

Med_ID: Medical ID (same as clinical). Units: -9999. Coding: Numeric

Age: Date of interview- Date of birth. Units: -9999. Coding: Numeric

Ethnicity: Participant ethnicity. Units: -9999. Coding: Text

ID_Gender: Gender. Units: -9999. Coding: {.00, male}{1.00, female}

ID_Education: What is the highest grade in school or college that you have passed or completed?. Units: -9999. Coding: 0-20

05_r6_LUMIPULSE_CSF_Abeta40: CSF Amyloid Beta 40_Lumipulse. Units: NA. Coding: NA

05_r6_LUMIPULSE_CSF_Abeta42: CSF Amyloid Beta 42_Lumipulse. Units: NA. Coding: NA

05_r6_LUMIPULSE_CSF_Total_Tau: CSF Total Tau_Lumipulse. Units: NA. Coding: NA

05_r6_LUMIPULSE_CSF_pTau181: CSF pTau181_Lumipulse. Units: NA. Coding: NA

05_r6_LUMIPULSE_CSF_NfL: CSF NfL_Lumipulse. Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: NA

r3_QTX_Plasma_Abeta40: Plasma Amyloid Beta 40_Quanterix. Units: NA. Coding: NA

r3_QTX_Plasma_Abeta42: Plasma Amyloid Beta 42_Quanterix. Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: NA

r3_QTX_Plasma_Total_Tau: Plasma Total Tau_Quanterix. Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: NA

r3_QTX_Plasma_IL_10: Plasma Interluekin 10 (IL 10)_Quanterix. Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg CV = 3.37190 Avg LLOD = 0.02866 Avg HLOD = 378 Concentration = pg/mL

r3_QTX_Plasma_pTau181: Plasma Phosphorylated Tau:Thr181 (pTau181)_Quanterix. Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg CV = 0.07065 Avg LLOD = 0.016 Avg HLOD = 200 Concentration = pg/mL

r3_QTX_Plasma_TNFalpha: Plasma Tumor Necrosis Factor Alpha (TNF-alpha)_Quanterix. Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg CV = 3.36104 Avg LLOD = 0.06597 Avg HLOD = 349 Concentration = pg/

r3_QTX_Plasma_IL_6: Plasma Interluekin 6 (IL 6)_Quanterix. Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg CV = 0.06030 Avg LLOD = 0.0055 Avg HLOD = 120 Concentration = pg/mL

r3_QTX_Plasma_NfL: Plasma Neurofilament Light Chain (NfL)_Quanterix. Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: NA

r3_QTX_Plasma_IL_5: Plasma Interluekin 5 (IL 5)_Quanterix. Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg CV = 0.07891 Avg LLOD = 0.0041 Avg HLOD = 12 Concentration = pg/mL

r3_MSD_Serum_CRP: Serum C-reactive protein (CRP)_Meso Scale Discovery. Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg CV = 2.66922 Avg LLOD = 3.28241 Avg HLOD = 196000 Concentration = pg/mL

r3_MSD_Serum_FABP3: Serum Fatty Acid Binding Protein 3 (FABP3)_Meso Scale Discovery. Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: NA

r3_MSD_Serum_IL_10: Serum Interleukin 10 (IL 10)_Meso Scale Discovery. Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg CV = 3.37190 Avg LLOD = 0.02866 Avg HLOD = 378 Concentration = pg/mL

r3_MSD_Serum_IL_6: Serum Interleukin 6 (IL 6)_Meso Scale Discovery. Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg CV = 6.69924 Avg LLOD = 0.05645 Avg HLOD = 742 Concentration = pg/mL

r3_MSD_Serum_PPY: Serum Pancreatic Polypeptide (PPY)_ Meso Scale Discovery. Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg CV = 5.08873 Avg LLOD = 150.14887 Avg HLOD = 16500 Concentration = pg/mL

r3_MSD_Serum_sICAM_1: Serum Intercellular Adhesion Molecule 1 (ICAM-1/CD54)_ Meso Scale Discovery. Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg CV = 6.31391 Avg LLOD = 2.60300 Avg HLOD = 50100 Concentration = pg/mL

r3_MSD_Serum_sVCAM_1: Serum Vascular Cell Adhesion Molecule 1 (VCAM-1/CD106)_Meso Scale Discovery. Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg CV = 3.87879 Avg LLOD = 7.95490 Avg HLOD = 47900 Concentration = pg/mL

r3_MSD_Serum_TNFalpha: Serum Human Tumor Necrosis Factor Alpha (TNF-alpha)_Meso Scale Discovery. Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: NA

r2_QTX_Plasma_Abeta40: Plasma Amyloid Beta 40_Quanterix. Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: NA

r2_QTX_Plasma_Abeta42: Plasma Amyloid Beta 42_Quanterix. Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg CV = 0.043 Avg LLOD = 0.045 Avg HLOD = 240 Concentration = pg/mL

r2_QTX_Plasma_Total_Tau: Plasma Total Tau_Quanterix. Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg CV = 0.061 Avg LLOD = 0.019 Avg HLOD = 400 Concentration = pg/mL

r2_QTX_Plasma_NfL: Plasma Neurofilament Light Chain (NfL)_Quanterix. Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg CV = 0.038 Avg LLOD = 0.038 Avg HLOD = 1800 Concentration = pg/mL

r3_MSD_Plasma_PYY: Plasma Peptide YY (PYY)_Meso Scale Discovery. Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: NA

r3_MSD_Plasma_GLP_1: Plasma Glucagon-Like-Peptide-1 (GLP-1)_Meso Scale Discovery. Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: NA

r3_MSD_Plasma_Glucagon: Plasma Glucagon_Meso Scale Discovery. Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: NA

r3_MSD_Plasma_Insulin: Plasma Insulin_Meso Scale Discovery. Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: NA

r5_QTX_Plasma_Abeta40: Plasma Amyloid Beta 40_Quanterix. Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: NA

r5_QTX_Plasma_Abeta42: Plasma Amyloid Beta 42_Quanterix. Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: NA

r5_QTX_Plasma_NfL: Plasma Neurofilament Light Chain (NfL)_Quanterix. Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: NA

r5_QTX_Plasma_pTau181: Plasma Phosphorylated Tau:Thr181 (pTau181)_Quanterix. Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: NA

r5_QTX_Plasma_Total_Tau: Plasma Total Tau_Quanterix. Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: NA

r5_QTX_Plasma_GFAP: Plasma Glial Fibrillary Acidic Protein (GFAP)_Quanterix. Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: NA

r5_LUM(Pro)_Plasma_Abeta40: Plasma Amyloid Beta 40_Luminex (Procarta). Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg Cv = -777777 Avg LLOD = 451 Avg HLOD = 1847000 Concentration = pg/mL

r5_LUM(Pro)_Plasma_Abeta42: Plasma Amyloid Beta 42_Luminex (Procarta). Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg Cv = 56.19 Avg LLOD = 0.42 Avg HLOD = 1700 Concentration = pg/mL

r5_LUM(Pro)_Plasma_CoV-229E Ig Total: Plasma Human Coronavirus 229E Immunoglobulin Total (CoV-229 Ig Total)_Luminex (Procarta). Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg Cv = 10.72 Avg LLOD = 0.15 Avg HLOD = 600 Concentration = U/mL

r5_LUM(Pro)_Plasma_CoV-HKU1 Ig Total: Plasma Human Coronavirus HKU1 Immunoglobulin Total (CoV-HKU1 Ig Total)_Luminex (Procarta). Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg Cv = 8.85 Avg LLOD = 0.12 Avg HLOD = 500 Concentration = U/mL

r5_LUM(Pro)_Plasma_CoV-NL63 Ig Total: Plasma Human Coronavirus NL63 Immunoglobulin Total (CoV-NL63 Ig Total)_Luminex (Procarta). Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg Cv = 7.03 Avg LLOD = 0.13 Avg HLOD = 550 Concentration = U/mL

r5_LUM(Pro)_Plasma_CoV-OC43 Ig Total: Plasma Human Coronavirus OC43 Immunoglobulin Total (CoV-OC43 Ig Total)_Luminex (Procarta). Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg Cv = 7.30 Avg LLOD = 0.12 Avg HLOD = 500 Concentration = U/mL

r5_LUM(Pro)_Plasma_GFAP: Plasma Glial Fibrillary Acidic Protein (GFAP)_Luminex (Procarta). Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg Cv = 91.69 Avg LLOD = 9.18 Avg HLOD = 37600 Concentration = pg/mL

r5_LUM(Pro)_Plasma_IgA: Plasma Immunoglobulin A (IgA)_Luminex (Procarta). Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg Cv = 14.36 Avg LLOD = 686 Avg HLOD = 500000 Concentration = pg/mL

r5_LUM(Pro)_Plasma_IgE: Plasma Immunoglobulin E (IgE)_Luminex (Procarta). Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg Cv = 18.87 Avg LLOD = 7.27 Avg HLOD = 5300 Concentration = pg/mL

r5_LUM(Pro)_Plasma_IgG1: Plasma Immunoglobulin G1 (IgG1)_Luminex (Procarta). Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg Cv = 15.43 Avg LLOD = 7545 Avg HLOD = 1833333 Concentration = pg/mL

r5_LUM(Pro)_Plasma_IgG2: Plasma Immunoglobulin G2 (IgG2)_Luminex (Procarta). Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg Cv = 20.53 Avg LLOD = 15089 Avg HLOD = 11000000 Concentration = pg/mL

r5_LUM(Pro)_Plasma_IgG3: Plasma Immunoglobulin G3 (IgG3)_Luminex (Procarta). Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg Cv = 7.26 Avg LLOD = 579 Avg HLOD = 140633 Concentration = pg/mL

r5_LUM(Pro)_Plasma_IgG4: Plasma Immunoglobulin G4 (IgG4)_Luminex (Procarta). Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg Cv = 23.32 Avg LLOD = 427 Avg HLOD = 311200 Concentration = pg/mL

r5_LUM(Pro)_Plasma_IgM: Plasma Immunoglobulin M (IgM)_Luminex (Procarta). Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg Cv = 36.06 Avg LLOD = 7267 Avg HLOD = 5298000 Concentration = pg/mL

r5_LUM(Pro)_Plasma_NCAM_1: Plasma Neural Cell Adhesion Molecule 1 (NCAM-1/CD56)_Luminex (Procarta). Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg Cv = 3.80 Avg LLOD = 54 Avg HLOD = 221900 Concentration = pg/mL

r5_LUM(Pro)_Plasma_Neurogranin (NRGN): Plasma Neurogranin (NRGN)_Luminex (Procarta). Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg Cv = 4.39 Avg LLOD = 10 Avg HLOD = 41600 Concentration = pg/mL

r5_LUM(Pro)_Plasma_NF-H: Plasma Neurofilament Heavy (NF-H)_Luminex (Procarta). Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg Cv = 73.01 Avg LLOD = 27 Avg HLOD = 27750 Concentration = pg/mL

r5_LUM(Pro)_Plasma_SARS S1 Ig Total: Plasma Severe Acute Respiratory Syndrome Spike Protein S1 Immunoglobulin Total (SARS S1 Ig Total)_Luminex (Procarta). Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg Cv = 8.07Avg LLOD = 0.073 Avg HLOD = 300 Concentration = U/mL

r5_LUM(Pro)_Plasma_SARS-CoV2 Nucleocapsid Ig Total: Plasma Severe Acute Respiratory Syndrome Coronavirus 2 Nucleocapsid Immunoglobulin Total (SARS-CoV2 Nucleocapsid Ig Total)_Luminex (Procarta). Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg Cv = 3.74 Avg LLOD = 0.11 Avg HLOD = 450 Concentration = U/mL

r5_LUM(Pro)_Plasma_SARS-CoV2 RBD Ig Total: Plasma Severe Acute Respiratory Syndrome Coronavirus 2 Receptor-Binding Domain Immunoglobulin Total (SARS-CoV2 RBD Ig Total)_Luminex (Procarta). Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg Cv = 7.6 Avg LLOD = 0.097 Avg HLOD = 400 Concentration = U/mL

r5_LUM(Pro)_Plasma_SARS-CoV2 S1 protein: Plasma Severe Acute Respiratory Syndrome Coronavirus 2 Spike Protein S1 (SARS-CoV2 S1)_Luminex (Procarta). Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg Cv = 7.60 Avg LLOD = 0.085 Avg HLOD = 350 Concentration = U/mL

r5_LUM(Pro)_Plasma_SARS-CoV2 Spike (trimer) Ig Tot: Plasma Severe Acute Respiratory Syndrome Coronavirus 2 Spike (trimer) Immunoglobulin Total (SARS-CoV2 Spike (trimer) Ig Total_Luminex (Procarta). Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg Cv = 6.745 Avg LLOD = 0.097 Avg HLOD = 400 Concentration = U/mL

r5_LUM(Pro)_Plasma_SARS-CoV2 Spike B.1.1.7: Plasma Severe Acute Respiratory Syndrome Coronavirus 2 Alpha Variant (SARS-CoV2 Spike B.1.1.7)_Luminex (Procarta). Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg Cv = 8.92 Avg LLOD = 0.11 Avg HLOD = 450 Concentration = U/mL

r5_LUM(Pro)_Plasma_SARS-CoV2 Spike B.1.351: Plasma Severe Acute Respiratory Syndrome Coronavirus 2 Beta Variant (SARS-CoV2 Spike B.1.351)_Luminex (Procarta). Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg Cv = 9.42 Avg LLOD = 0.11 Avg HLOD = 450 Concentration = U/mL

r5_LUM(Pro)_Plasma_SARS-CoV2 Spike B.1.617.2: Plasma Severe Acute Respiratory Syndrome Coronavirus 2 Delta Variant (SARS-CoV2 Spike B.1.617.2)_Luminex (Procarta). Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg Cv = 8.18 Avg LLOD = 0.085 Avg HLOD =350 Concentration = U/mL

r5_LUM(Pro)_Plasma_SARS-CoV2 Spike P.1: Plasma Severe Acute Respiratory Syndrome Coronavirus 2 Gamma Variant (SARS-CoV2 Spike P.1)_Luminex (Procarta). Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg Cv = 6.57 Avg LLOD = Avg HLOD = 400 Concentration = U/mL

r5_LUM(Pro)_Plasma_pTau181: Plasma Phosphorylated Tau:Thr181 (pTau181)_Luminex (Procarta). Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg Cv = 141.42 Avg LLOD = 1.95 Avg HLOD =2000 Concentration = pg/mL

r5_LUM(Pro)_Plasma_Total_Tau: Plasma Total Tau_Luminex (Procarta). Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg Cv = 30.01 Avg LLOD = 20 Avg HLOD = 80000 Concentration = pg/mL

r5_LUM(Pro)_Plasma_TDP-43: Plasma TAR DNA-Binding Protein 43 (TDP-43)_Luminex (Procarta). Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg Cv = -777777 Avg LLOD = 96 Avg HLOD = 393200 Concentration = pg/mL

r5_LUM(Pro)_Plasma_YKL-40 (CHI3L1): Plasma Chitinase-3-Like Protein 1 (YKL-40/CHI3L1)_Luminex (Procarta). Units: Below detection limit -999999 ; Above detection limit -888888 ; Missing -777777. Coding: Avg Cv = 5.76 Avg LLOD = 32 Avg HLOD = 332225 Concentration = pg/mL

HOMA_IR: Homeostatic Model Assessment for Insulin Resistance. A quantitative assessment of contributions of insulin resistance and deficient beta-cell function to the fasting hyperglycemia.. Units: -9999. Coding: A computed variable using BW_Glucose and BW_Insulin.

Visit_ID: Visit Id number. Units: NA. Coding: Numeric

Med_ID: Medical Id number. Note: This is not an MRN and is not identifiable.. Units: NA. Coding: Numeric

Age: date of interview- date of birth. Units: (-9999 = N/A), (-8888 =Refused). Coding: Numeric

Ethnicity: Participant ethnicity. Units: (-9999 = N/A), (-8888 =Refused). Coding: Text

ID_Gender: Gender. Units: (-9999 = N/A), (-8888 =Refused). Coding: {.00, male}{1.00, female}

ID_Education: What is the highest grade in school or college that you have passed or completed?. Units: (-9999 = N/A), (-8888 =Refused). Coding: 0-20

Visit_ID: Visit Number. Units: -9999. Coding: Numeric

Med_ID: Medical ID (same as clinical data). Units: -9999. Coding: Numeric

Age: Date of interview- Date of birth. Units: -9999. Coding: Numeric

Ethnicity: Participant ethnicity. Units: -9999. Coding: Text

ID_Gender: Gender. Units: -9999. Coding: {.00, male}{1.00, female}

ID_Education: What is the highest grade in school or college that you have passed or completed?. Units: -9999. Coding: 0-20

Visit_ID: Visit number. Units: NA. Coding: Numeric

Med_ID: Medical ID (same as clinical). Units: NA. Coding: Numeric

Age: Date of interview- Date of birth. Units: NA. Coding: Numeric

Ethnicity: Participant ethnicity. Units: NA. Coding: Text

ID_Gender: Gender. Units: NA. Coding: {.00, male}{1.00, female}

ID_Education: What is the highest grade in school or college that you have passed or completed?. Units: NA. Coding: 0-20

Examples

if (FALSE) { # \dontrun{
browseVignettes('HABSHD')
} # }