HD_Genomics

HABS-HD dataset.

Usage

data(HD_Genomics)

Format

A data frame with 3343 rows and 10 variables.

Details

Visit_ID: Visit number. Units: -9999. Coding: Numeric

Med_ID: Medical ID (same as clinical). Units: -9999. Coding: Numeric

Age: Date of interview- Date of birth. Units: -9999. Coding: Numeric

Ethnicity: Participant ethnicity. Units: -9999. Coding: Text

ID_Gender: Gender. Units: -9999. Coding: {.00, male}{1.00, female}

ID_Education: What is the highest grade in school or college that you have passed or completed?. Units: -9999. Coding: 0-20

Visit_ID: Visit Id number. Units: NA. Coding: Numeric

Med_ID: Medical Id number. Note: This is not an MRN and is not identifiable.. Units: NA. Coding: Numeric

Age: date of interview- date of birth. Units: (-9999 = N/A), (-8888 =Refused). Coding: Numeric

Ethnicity: Participant ethnicity. Units: (-9999 = N/A), (-8888 =Refused). Coding: Text

ID_Gender: Gender. Units: (-9999 = N/A), (-8888 =Refused). Coding: {.00, male}{1.00, female}

ID_Education: What is the highest grade in school or college that you have passed or completed?. Units: (-9999 = N/A), (-8888 =Refused). Coding: 0-20

Visit_ID: Visit Number. Units: -9999. Coding: Numeric

Med_ID: Medical ID (same as clinical data). Units: -9999. Coding: Numeric

Age: Date of interview- Date of birth. Units: -9999. Coding: Numeric

Ethnicity: Participant ethnicity. Units: -9999. Coding: Text

ID_Gender: Gender. Units: -9999. Coding: {.00, male}{1.00, female}

ID_Education: What is the highest grade in school or college that you have passed or completed?. Units: -9999. Coding: 0-20

APOE4_rs429358: rs429358 alleles, Implicated SNP in non-Hispanic White AD (Kunkle et al., 2019); nearest gene is APOE. Units: -9999. Coding: SNP Genotypes

APOE4_rs7412: rs7412 alleles. Units: -9999. Coding: SNP Genotypes

APOE4_Genotype: APOE SNP GENOTYPE. Units: -9999. Coding: SNP Genotypes

APOE4_Positivity: {0, Absent}{1, Present). Units: -9999. Coding: SNP Genotypes -{0: Absent}{1: Present)

Visit_ID: Visit number. Units: NA. Coding: Numeric

Med_ID: Medical ID (same as clinical). Units: NA. Coding: Numeric

Age: Date of interview- Date of birth. Units: NA. Coding: Numeric

Ethnicity: Participant ethnicity. Units: NA. Coding: Text

ID_Gender: Gender. Units: NA. Coding: {.00, male}{1.00, female}

ID_Education: What is the highest grade in school or college that you have passed or completed?. Units: NA. Coding: 0-20

Examples

if (FALSE) { # \dontrun{
browseVignettes('HABSHD')
} # }