HD_Imaging

HABS-HD dataset.

Usage

data(HD_Imaging)

Format

A data frame with 5962 rows and 287 variables.

Details

Visit_ID: Visit number. Units: -9999. Coding: Numeric

Med_ID: Medical ID (same as clinical). Units: -9999. Coding: Numeric

Age: Date of interview- Date of birth. Units: -9999. Coding: Numeric

Ethnicity: Participant ethnicity. Units: -9999. Coding: Text

ID_Gender: Gender. Units: -9999. Coding: {.00, male}{1.00, female}

ID_Education: What is the highest grade in school or college that you have passed or completed?. Units: -9999. Coding: 0-20

Visit_ID: Visit Id number. Units: NA. Coding: Numeric

Med_ID: Medical Id number. Note: This is not an MRN and is not identifiable.. Units: NA. Coding: Numeric

Age: date of interview- date of birth. Units: (-9999 = N/A), (-8888 =Refused). Coding: Numeric

Ethnicity: Participant ethnicity. Units: (-9999 = N/A), (-8888 =Refused). Coding: Text

ID_Gender: Gender. Units: (-9999 = N/A), (-8888 =Refused). Coding: {.00, male}{1.00, female}

ID_Education: What is the highest grade in school or college that you have passed or completed?. Units: (-9999 = N/A), (-8888 =Refused). Coding: 0-20

Visit_ID: Visit Number. Units: -9999. Coding: Numeric

Med_ID: Medical ID (same as clinical data). Units: -9999. Coding: Numeric

Age: Date of interview- Date of birth. Units: -9999. Coding: Numeric

Ethnicity: Participant ethnicity. Units: -9999. Coding: Text

ID_Gender: Gender. Units: -9999. Coding: {.00, male}{1.00, female}

ID_Education: What is the highest grade in school or college that you have passed or completed?. Units: -9999. Coding: 0-20

Visit_ID: Visit number. Units: NA. Coding: Numeric

Med_ID: Medical ID (same as clinical). Units: NA. Coding: Numeric

Age: Date of interview- Date of birth. Units: NA. Coding: Numeric

Ethnicity: Participant ethnicity. Units: NA. Coding: Text

ID_Gender: Gender. Units: NA. Coding: {.00, male}{1.00, female}

ID_Education: What is the highest grade in school or college that you have passed or completed?. Units: NA. Coding: 0-20

01_MRI_Scanner: Scan Protocol - differentiates the named Siemens scanners used to acquire subject data. Units: NA. Coding: Skyra, Vida1, Vida2

01_MRI_Time_point: Time point for the subject for any given scan. Scan1, Scan2, Scan3 etc.. Units: NA. Coding: Scan1 / Scan2 / Scan3 /Scan4

01_MRI_Date_Days_Differential: Number of days between MRI scan and Interview date. Units: NA. Coding: Numeric

01_L_bankssts_thickavg: Banks of the superior temporal sulcus thickness determined on a T1-weighted volume scan - left. Units: NA. Coding: NA

01_L_caudalanteriorcingulate_thickavg: Caudal anterior cingulate thickness determined on a T1-weighted volume scan - left. Units: NA. Coding: NA

01_L_caudalmiddlefrontal_thickavg: Caudal middle frontal thickness determined on a T1-weighted volume scan - left. Units: NA. Coding: NA

01_L_cuneus_thickavg: Cuneus thickness determined on a T1-weighted volume scan - left. Units: NA. Coding: NA

01_L_entorhinal_thickavg: Entorhinal thickness determined on a T1-weighted volume scann - left. Units: NA. Coding: NA

01_L_fusiform_thickavg: Fusiform thickness determined on a T1-weighted volume scan - left. Units: NA. Coding: NA

01_L_inferiorparietal_thickavg: Inferior parietal thickness determined on a T1-weighted volume scan - left. Units: NA. Coding: NA

01_L_inferiortemporal_thickavg: Inferior temporal thickness determined on a T1-weighted volume scan - left. Units: NA. Coding: NA

01_L_isthmuscingulate_thickavg: Isthmus cingulate thickness determined on a T1-weighted volume scan - left. Units: NA. Coding: NA

01_L_lateraloccipital_thickavg: Lateral occipital thickness determined on a T1-weighted volume scan - left. Units: NA. Coding: NA

01_L_lateralorbitofrontal_thickavg: Lateral orbitofrontal thickness determined on a T1-weighted volume scan - left. Units: NA. Coding: NA

01_L_lingual_thickavg: Lingual thickness determined on a T1-weighted volume scan - left. Units: NA. Coding: NA

01_L_medialorbitofrontal_thickavg: Medial orbitofrontal thickness determined on a T1-weighted volume scan - left. Units: NA. Coding: NA

01_L_middletemporal_thickavg: Middle temporal thickness determined on a T1-weighted volume scan - left. Units: NA. Coding: NA

01_L_parahippocampal_thickavg: Parahippocampal thickness determined on a T1-weighted volume scan - left. Units: NA. Coding: NA

01_L_paracentral_thickavg: Paracentral thickness determined on a T1-weighted volume scan - left. Units: NA. Coding: NA

01_L_parsopercularis_thickavg: Parsopercularis thickness - determined on a T1-weighted volume scan- left. Units: NA. Coding: NA

01_L_parsorbitalis_thickavg: Parsorbitalis thickness determined on a T1-weighted volume scan- left. Units: NA. Coding: NA

01_L_parstriangularis_thickavg: Parstriangularis thickness determined on a T1-weighted volume scan- left. Units: NA. Coding: NA

01_L_pericalcarine_thickavg: Pericalcarine thickness determined on a T1-weighted volume scan - left. Units: NA. Coding: NA

01_L_postcentral_thickavg: Postcentral thickness determined on a T1-weighted volume scan - left. Units: NA. Coding: NA

01_L_posteriorcingulate_thickavg: Posterior cingulate thickness determined on a T1-weighted volume scan - left. Units: NA. Coding: NA

01_L_precentral_thickavg: Precentral thickness determined on a T1-weighted volume scan - left. Units: NA. Coding: NA

01_L_precuneus_thickavg: Precuneus thickness determined on a T1-weighted volume scan - left. Units: NA. Coding: NA

01_L_rostralanteriorcingulate_thickavg: Rostral anterior cingulate thickness determined on a T1-weighted volume scan - left. Units: NA. Coding: NA

01_L_rostralmiddlefrontal_thickavg: Rostral middle frontal thickness determined on a T1-weighted volume scan - left. Units: NA. Coding: NA

01_L_superiorfrontal_thickavg: Superior frontal thickness determined on a T1-weighted volume scan - left. Units: NA. Coding: NA

01_L_superiorparietal_thickavg: Superior parietal thickness determined on a T1-weighted volume scan - left. Units: NA. Coding: NA

01_L_superiortemporal_thickavg: Superior temporal thickness determined on a T1-weighted volume scan - left. Units: NA. Coding: NA

01_L_supramarginal_thickavg: Supramarginal thickness determined on a T1-weighted volume scan - left. Units: NA. Coding: NA

01_L_frontalpole_thickavg: Frontal pole thickness determined on a T1-weighted volume scan - left. Units: NA. Coding: NA

01_L_temporalpole_thickavg: Temporal pole thickness determined on a T1-weighted volume scan - left. Units: NA. Coding: NA

01_L_transversetemporal_thickavg: Transverse temporal thickness determined on a T1-weighted volume scan - left. Units: NA. Coding: NA

01_L_insula_thickavg: Insula thickness determined on a T1-weighted volume scan - left. Units: NA. Coding: NA

01_R_bankssts_thickavg: Banks of the superior temporal sulcus thickness determined on a T1-weighted volume scan - right. Units: NA. Coding: NA

01_R_caudalanteriorcingulate_thickavg: Caudal anterior cingulate thickness determined on a T1-weighted volume scan - right. Units: NA. Coding: NA

01_R_caudalmiddlefrontal_thickavg: Caudal middle frontal thickness determined on a T1-weighted volume scan - right. Units: NA. Coding: NA

01_R_cuneus_thickavg: Cuneus thickness determined on a T1-weighted volume scan - right. Units: NA. Coding: NA

01_R_entorhinal_thickavg: Entorhinal thickness determined on a T1-weighted volume scan - right. Units: NA. Coding: NA

01_R_fusiform_thickavg: Fusiform thickness determined on a T1-weighted volume scan - right. Units: NA. Coding: NA

01_R_inferiorparietal_thickavg: Inferior parietal thickness determined on a T1-weighted volume scan - right. Units: NA. Coding: NA

01_R_inferiortemporal_thickavg: Inferior temporal thickness determined on a T1-weighted volume scan - right. Units: NA. Coding: NA

01_R_isthmuscingulate_thickavg: Isthmus cingulate thickness determined on a T1-weighted volume scan - right. Units: NA. Coding: NA

01_R_lateraloccipital_thickavg: Lateral occipital thickness determined on a T1-weighted volume scan - right. Units: NA. Coding: NA

01_R_lateralorbitofrontal_thickavg: Lateral orbitofrontal thickness determined on a T1-weighted volume scan - right. Units: NA. Coding: NA

01_R_lingual_thickavg: Lingual thickness determined on a T1-weighted volume scan - right. Units: NA. Coding: NA

01_R_medialorbitofrontal_thickavg: Medial orbitofrontal thickness determined on a T1-weighted volume scan - right. Units: NA. Coding: NA

01_R_middletemporal_thickavg: Middle temporal thickness determined on a T1-weighted volume scan - right. Units: NA. Coding: NA

01_R_parahippocampal_thickavg: Parahippocampal thickness determined on a T1-weighted volume scan - right. Units: NA. Coding: NA

01_R_paracentral_thickavg: Paracentral thickness determined on a T1-weighted volume scan - right. Units: NA. Coding: NA

01_R_parsopercularis_thickavg: Parsopercularis thickness determined on a T1-weighted volume scan - right. Units: NA. Coding: NA

01_R_parsorbitalis_thickavg: Parsorbitalis thickness determined on a T1-weighted volume scan - right. Units: NA. Coding: NA

01_R_parstriangularis_thickavg: Parstriangularis thickness determined on a T1-weighted volume scan - right. Units: NA. Coding: NA

01_R_pericalcarine_thickavg: Pericalcarine thickness determined on a T1-weighted volume scan - right. Units: NA. Coding: NA

01_R_postcentral_thickavg: Postcentral thickness determined on a T1-weighted volume scan - right. Units: NA. Coding: NA

01_R_posteriorcingulate_thickavg: Posterior cingulate thickness determined on a T1-weighted volume scan - right. Units: NA. Coding: NA

01_R_precentral_thickavg: Precentral thickness determined on a T1-weighted volume scan - right. Units: NA. Coding: NA

01_R_precuneus_thickavg: Precuneus thickness determined on a T1-weighted volume scan - right. Units: NA. Coding: NA

01_R_rostralanteriorcingulate_thickavg: Rostral anterior cingulate thickness determined on a T1-weighted volume scan - right. Units: NA. Coding: NA

01_R_rostralmiddlefrontal_thickavg: Rostral middle frontal thickness determined on a T1-weighted volume scan - right. Units: NA. Coding: NA

01_R_superiorfrontal_thickavg: Superior frontal thickness determined on a T1-weighted volume scan - right. Units: NA. Coding: NA

01_R_superiorparietal_thickavg: Superior parietal thickness determined on a T1-weighted volume scan - right. Units: NA. Coding: NA

01_R_superiortemporal_thickavg: Superior temporal thickness determined on a T1-weighted volume scan - right. Units: NA. Coding: NA

01_R_supramarginal_thickavg: Supramarginal thickness determined on a T1-weighted volume scan - right. Units: NA. Coding: NA

01_R_frontalpole_thickavg: Frontal pole thickness determined on a T1-weighted volume scan - right. Units: NA. Coding: NA

01_R_temporalpole_thickavg: Temporal pole thickness determined on a T1-weighted volume scan - right. Units: NA. Coding: NA

01_R_transversetemporal_thickavg: Transverse temporal thickness determined on a T1-weighted volume scan - right. Units: NA. Coding: NA

01_R_insula_thickavg: Insula thickness determined on a T1-weighted volume scan - right. Units: NA. Coding: NA

01_CA1_L_volume: CA1 of the hippocampus volume determined on a hippocampal high-resolution scan - left; please control for number of slices segmented Slices_seg and intracranial volume ICV_FS. Units: NA. Coding: NA

01_CA1_R_volume: CA1 of the hippocampus volume determined on a hippocampal high-resolution scan - right; please control for number of slices segmented Slices_seg and intracranial volume ICV_FS. Units: NA. Coding: NA

01_CA23DG_L_volume: CA2, 3, & dentate gyrus of the hippocampus volume determined on a hippocampal high-resolution scan - left; please control for number of slices segmented Slices_seg and intracranial volume ICV_FS. Units: NA. Coding: NA

01_CA23DG_R_volume: CA2, 3, & dentate gyrus of the hippocampus volume determined on a hippocampal high-resolution scan - right; please control for number of slices segmented Slices_seg and intracranial volume ICV_FS. Units: NA. Coding: NA

01_SUB_L_volume: Subiculum volume determined on a hippocampal high-resolution scan - left; please control for number of slices segmented Slices_seg and intracranial volume ICV_FS. Units: NA. Coding: NA

01_SUB_R_volume: Subiculum volume determined on a hippocampal high-resolution scan - right; please control for number of slices segmented Slices_seg and intracranial volume ICV_FS. Units: NA. Coding: NA

01_Slices_seg_L: Number of slices segmented on the hippocampal high resolution scans for the left hippocampus. For all analyses that include hippocampal subregion volumes derived from the hippocampal high resolution scans, number of slices segmented should be controlled for along with intracranial volume (ICV_FS).. Units: NA. Coding: NA

01_Slices_seg_R: Number of slices segmented on the hippocampal high resolution scans for the right hippocampus. For all analyses that include hippocampal subregion volumes derived from the hippocampal high resolution scans, number of slices segmented should be controlled for along with intracranial volume (ICV_FS).. Units: NA. Coding: NA

01_L_hippocampus: Whole hippocampal volume determined on a T1-weighted volume scan - left; control for intracranial volume (ICV_HD). Units: NA. Coding: NA

01_R_hippocampus: Whole hippocampal volume determined on a T1-weighted volume scan - right; control for intracranial volume (ICV_HD). Units: NA. Coding: NA

01_ICV_HD: Hippodeep- estimated intracranial volume determined on a T1-weighted volume scan. Please note that we previously used ICV from FreeSurfer. These two variables are slightly different and should not be mixed.. Units: NA. Coding: NA

01_CorticalThickness_MetaROI: Surface area-weighted average of cortical thickness in rightand left entorhinal cortex, fusiform, and inferior and middle temporal cortex. Units: NA. Coding: NA

01_BrainSegNotVent: FreeSurfer- brain segmentation volume without ventricles (lateral, inferior lateral, 3rd, 4th, 5th), CSF, and choroid plexus. Units: NA. Coding: NA

01_AB_FBB_Scanner: PET scanner: MCT20 (Siemens MCT 20 PET/CT) or Vision (Siemens BioGraph Vision 450 PET/CT). Units: NA. Coding: MCT20 / Vision

01_AB_FBB_Date_Days_Differential: Number of days between Amyloid scan and interview date. Units: NA. Coding: Numeric

01_AB_FBB_Time_point: Time point for the subject for any given scan. Scan1, Scan2, Scan3 etc.. Units: NA. Coding: Scan1 / Scan2 / Scan3 /Scan4

01_AB_FBB_Frontal_SUVR: Amyloid Frontal SUVR - ADNI2 regions for FBB; whole cerebellum reference; PET scanner should be considered in statistical analyses.. Units: NA. Coding: NA

01_AB_FBB_AntPostCingulate_SUVR: Amyloid Anterior/posterior cingulate SUVR - ADNI2 regions for FBB; whole cerebellum reference; PET scanner should be considered in statistical analyses.. Units: NA. Coding: NA

01_AB_FBB_LatParietal_SUVR: Amyloid Lateral Parietal SUVR - ADNI2 regions for FBB; whole cerebellum reference; PET scanner should be considered in statistical analyses.. Units: NA. Coding: NA

01_AB_FBB_LatTemporal_SUVR: Amyloid Lateral Temporal SUVR - ADNI2 regions for FBB; whole cerebellum reference. PET scanner should be considered in statistical analyses.. Units: NA. Coding: NA

01_AB_FBB_Global_SUVR: Amyloid Global SUVR - ADNI2 regions for FBB; whole cerebellum reference; PET scanner should be considered in statistical analyses.. Units: NA. Coding: NA

01_AB_FBB_AB_pos: Amyloid positivity as defined by ADNI3 for FBB using a whole cerebellum reference region; Global SUVR >1.08; PET scanner should be considered in statistical analyses.. Units: NA. Coding: NA

01_TAU_PI2620_Slice_thickness: Slice thickness for reeconstructed 18F-PI-2620 scans; 0 for standard slice thickness, 1 for alternative slice thickness. Units: NA. Coding: NA

01_TAU_PI2620_Scanner: PET scanner: MCT20 (Siemens MCT 20 PET/CT) or Vision (Siemens BioGraph Vision 450 PET/CT). Units: NA. Coding: Vision

01_TAU_PI2620_Time_point: Time point for the subject for any given scan. Scan1, Scan2, Scan3 etc.. Units: NA. Coding: Scan1 / Scan2 / Scan3 /Scan4

01_TAU_PI2620_Date_Days_Differential: Number of days between TAU scan and interview date. Units: NA. Coding: Numeric

01_TAU_PI2620_Medial_Temporal_SUVR: Medial Temporal SUVR - Mormino et al 2020 regions for 18F-PI-2620 (entorhinal cortex, hippocampus, amygdala); inferior cerebellar grey matter reference region. Units: NA. Coding: NA

01_TAU_PI2620_Posterior_Cingulate_SUVR: PosteriorCingulate SUVR - Mormino et al 2020 for 18F-PI-2620 (isthmus cingulate and posterior cingulate); inferior cerebellar grey matter reference region. Units: NA. Coding: NA

01_TAU_PI2620_Lateral_Parietal_SUVR: Lateral Pariental SUVR - Mormino et al 2020 for 18F-PI-2620 (interior parietal, superior parietal, and supramarginal); inferior cerebellar grey matter reference region. Units: NA. Coding: NA

01_TAU_PI2620_Median_Medial_Temporal_SUVR: Median Tau Medial Temporal SUVR - ADNI3 regions for 18F-PI-2620; inferior cerebellar grey matter reference region. Median values are less influenced by outlying data points (such as from off target binding) than mean values are.. Units: NA. Coding: NA

01_TAU_PI2620_Median_Posterior_Cingulate_SUVR: Median Tau Posterior Cingulate SUVR - ADNI3 regions for 18F-PI-2620; inferior cerebellar grey matter reference region. Median values are less influenced by outlying data points (such as from off target binding) than mean values are.. Units: NA. Coding: NA

01_TAU_PI2620_Median_Lateral_Parietal_SUVR: Median Tau Lateral Parietal SUVR - ADNI3 regions for 18F-PI-2620; inferior cerebellar grey matter reference region. Median values are less influenced by outlying data points (such as from off target binding in meninges) than mean values are.. Units: NA. Coding: NA

01_TAU_PI2620_Eroded_Lateral_Parietal_SUVR: Eroded Tau Lateral Parietal SUVR - ADNI3 regions for 18F-PI-2620; inferior cerebellar grey matter reference region. Mean value in an eroded lateral parietal region that is less likely to include off target binding from meninges.. Units: NA. Coding: NA

02_WMH_Scale: White Matter Hyperintensity Scale {0, No Significance}{1, Mild}{2, Moderate}{3, Severe}. Units: NA. Coding: NA

02_Microhem_Count: Hemosiderin count. Units: NA. Coding: NA

02_Microhem_Present: Hemosiderin present (yes = 1 / no = 0). Units: NA. Coding: {0:No, 1: Yes}

02_Lacune_Count: Lacune Count. Units: NA. Coding: NA

02_Lacune_Present: Lacune Present (yes = 1 / no = 0). Units: NA. Coding: {0:No, 1: Yes}

02_Calc_Count: NA. Units: NA. Coding: NA

02_Calc_Present: NA. Units: NA. Coding: NA

02_Consider_Exclusion: NA. Units: NA. Coding: NA

02_Clinical_Findings_MH_microhemorrhage_calc_calcification: NA. Units: NA. Coding: NA

02_Encephalomalacia_Favor_Stroke: NA. Units: NA. Coding: NA

02_Encephalomalacia_Favor_TBI: NA. Units: NA. Coding: NA

02_Encephalomalacia: NA. Units: NA. Coding: NA

02_Definite_Exclude: NA. Units: NA. Coding: NA

02_Suggest_Exclude: NA. Units: NA. Coding: NA

02_Clinical_Impression: NA. Units: NA. Coding: NA

02_Clinical_Findings: NA. Units: NA. Coding: NA

02_WMH_Volume_Raw: White Matter Hyperintensity Volume. Units: NA. Coding: NA

02_WMH_Volume_Log: NA. Units: NA. Coding: NA

03_AD_Corticospinal_tract_left: Axial Diffusivity Corticospinal tract left. Units: NA. Coding: NA

03_AD_Anterior_limb_of_internal_capsule_left: Axial Diffusivity Anterior limb of internal capsule left. Units: NA. Coding: NA

03_AD_Posterior_limb_of_internal_capsule_left: Axial Diffusivity Posterior limb of internal capsule left. Units: NA. Coding: NA

03_AD_Posterior_thalamic_radiation_left: Axial Diffusivity Posterior thalamic radiation left. Units: NA. Coding: NA

03_AD_Anterior_corona_radiata_left: Axial Diffusivity Anterior coronara radiatia left. Units: NA. Coding: NA

03_AD_Superior_corona_radiata_left: Axial Diffusivity Superior corona radiatia left. Units: NA. Coding: NA

03_AD_Posterior_corona_radiata_left: Axial Diffusivity Posterior corona radiatia left. Units: NA. Coding: NA

03_AD_Cingulum_cingulate_gyrus_left: Axial Diffusivity Cingulum cingulate gyrus left. Units: NA. Coding: NA

03_AD_Cingulum_hippocampus_left: Axial Diffusivity Cingulum hippocampus left. Units: NA. Coding: NA

03_AD_Superior_longitudinal_fasciculus_left: Axial Diffusivity Superior longitudinal fasciculus left. Units: NA. Coding: NA

03_AD_Superior_fronto_occipital_fasciculus_left: Axial Diffusivity Superior fronto occipital fasciculus left. Units: NA. Coding: NA

03_AD_Inferior_fronto_occipital_fasciculus_left: Axial Diffusivity Inferior fronto occipital fasciculus left. Units: NA. Coding: NA

03_AD_Uncinate_fasciculus_left: Axial Diffusivity Uncinate fasciculus left. Units: NA. Coding: NA

03_AD_Genu_of_corpus_callosum_left: Axial Diffusivity Genu of corpus callosum left. Units: NA. Coding: NA

03_AD_Body_of_corpus_callosum_left: Axial Diffusivity Body of corpus callosum left. Units: NA. Coding: NA

03_AD_Splenium_of_corpus_callosum_left: Axial Diffusivity Splenium of corpus callosum left. Units: NA. Coding: NA

03_AD_Corticospinal_tract_right: Axial Diffusivity Corticospinal tract right. Units: NA. Coding: NA

03_AD_Anterior_limb_of_internal_capsule_right: Axial Diffusivity Anterior limb of internal capsule right. Units: NA. Coding: NA

03_AD_Posterior_limb_of_internal_capsule_right: Axial Diffusivity Posterior limb of internal capsule right. Units: NA. Coding: NA

03_AD_Posterior_thalamic_radiation_include_optic_radiation_right: Axial Diffusivity Posterior thalamic radiation include optic radiation right. Units: NA. Coding: NA

03_AD_Anterior_corona_radiata_right: Axial Diffusivity Anterior corona radiata right. Units: NA. Coding: NA

03_AD_Superior_corona_radiata_right: Axial Diffusivity Superior corona radiata right. Units: NA. Coding: NA

03_AD_Posterior_corona_radiata_right: Axial Diffusivity Posterior corona radiata right. Units: NA. Coding: NA

03_AD_Cingulum_cingulate_gyrus_right: Axial Diffusivity Cingulum cingulate gyrus right. Units: NA. Coding: NA

03_AD_Cingulum_hippocampus_right: Axial Diffusivity Cingulum hippocampus right. Units: NA. Coding: NA

03_AD_Superior_longitudinal_fasciculus_right: Axial Diffusivity Superior longitudinal fasciculus right. Units: NA. Coding: NA

03_AD_Superior_fronto_occipital_fasciculus_right: Axial Diffusivity Superior fronto occipital fasciculus right. Units: NA. Coding: NA

03_AD_Inferior_fronto_occipital_fasciculus_right: Axial Diffusivity Inferior fronto occipital fasciculus right. Units: NA. Coding: NA

03_AD_Uncinate_fasciculus_right: Axial Diffusivity Uncinate fasciculus right. Units: NA. Coding: NA

03_AD_Genu_of_corpus_callosum_right: Axial Diffusivity Genu of corpus callosum right. Units: NA. Coding: NA

03_AD_Body_of_corpus_callosum_right: Axial Diffusivity Body of corpus callosum right. Units: NA. Coding: NA

03_AD_Splenium_of_corpus_callosum_right: Axial Diffusivity Splenium of corpus callosum right. Units: NA. Coding: NA

03_FA_Corticospinal_tract_left: Fractional Anisotropy Corticospinal tract left. Units: NA. Coding: Fiber bundles

03_FA_Anterior_limb_of_internal_capsule_left: Fractional Anisotropy Anterior limb of internal capsule left. Units: NA. Coding: Fiber bundles

03_FA_Posterior_limb_of_internal_capsule_left: Fractional Anisotropy Posterior limb of internal capsule left. Units: NA. Coding: Fiber bundles

03_FA_Posterior_thalamic_radiation_left: Fractional Anisotropy Posterior thalamic radiation left. Units: NA. Coding: Fiber bundles

03_FA_Anterior_corona_radiata_left: Fractional Anisotropy Anterior corona radiata left. Units: NA. Coding: Fiber bundles

03_FA_Superior_corona_radiata_left: Fractional Anisotropy Superior corona radiata left. Units: NA. Coding: Fiber bundles

03_FA_Posterior_corona_radiata_left: Fractional Anisotropy Posterior corona radiata left. Units: NA. Coding: Fiber bundles

03_FA_Cingulum_cingulate_gyrus_left: Fractional Anisotropy Cingulum cingulate gyrus left. Units: NA. Coding: Fiber bundles

03_FA_Cingulum_hippocampus_left: Fractional Anisotropy Cingulum hippocampus left. Units: NA. Coding: Fiber bundles

03_FA_Superior_longitudinal_fasciculus_left: Fractional Anisotropy Superior longitudinal fasciculus left. Units: NA. Coding: Fiber bundles

03_FA_Superior_fronto_occipital_fasciculus_left: Fractional Anisotropy Superior fronto occipital fasciculus left. Units: NA. Coding: Fiber bundles

03_FA_Inferior_fronto_occipital_fasciculus_left: Fractional Anisotropy Inferior fronto occipital fasciculus left. Units: NA. Coding: Fiber bundles

03_FA_Uncinate_fasciculus_left: Fractional Anisotropy Uncinate fasciculus left. Units: NA. Coding: Fiber bundles

03_FA_Genu_of_corpus_callosum_left: Fractional Anisotropy Genu of corpus callosum left. Units: NA. Coding: Fiber bundles

03_FA_Body_of_corpus_callosum_left: Fractional Anisotropy Body of corpus callosum left. Units: NA. Coding: Fiber bundles

03_FA_Splenium_of_corpus_callosum_left: Fractional Anisotropy Splenium of corpus callosum left. Units: NA. Coding: Fiber bundles

03_FA_Corticospinal_tract_right: Fractional Anisotropy Corticospinal tract right. Units: NA. Coding: Fiber bundles

03_FA_Anterior_limb_of_internal_capsule_right: Fractional Anisotropy Anterior limb of internal capsule right. Units: NA. Coding: Fiber bundles

03_FA_Posterior_limb_of_internal_capsule_right: Fractional Anisotropy Posterior limb of internal capsule right. Units: NA. Coding: Fiber bundles

03_FA_Posterior_thalamic_radiation_include_optic_radiation_right: Fractional Anisotropy Posterior thalamic radiation include optic radiation right. Units: NA. Coding: Fiber bundles

03_FA_Anterior_corona_radiata_right: Fractional Anisotropy Anterior corona radiata right. Units: NA. Coding: Fiber bundles

03_FA_Superior_corona_radiata_right: Fractional Anisotropy Superior corona radiata right. Units: NA. Coding: Fiber bundles

03_FA_Posterior_corona_radiata_right: Fractional Anisotropy Posterior corona radiata right. Units: NA. Coding: Fiber bundles

03_FA_Cingulum_cingulate_gyrus_right: Fractional Anisotropy Cingulum cingulate gyrus right. Units: NA. Coding: Fiber bundles

03_FA_Cingulum_hippocampus_right: Fractional Anisotropy Cingulum hippocampus right. Units: NA. Coding: Fiber bundles

03_FA_Superior_longitudinal_fasciculus_right: Fractional Anisotropy Superior longitudinal fasciculus right. Units: NA. Coding: Fiber bundles

03_FA_Superior_fronto_occipital_fasciculus_right: Fractional Anisotropy Superior fronto occipital fasciculus right. Units: NA. Coding: Fiber bundles

03_FA_Inferior_fronto_occipital_fasciculus_right: Fractional Anisotropy Inferior fronto occipital fasciculus right. Units: NA. Coding: Fiber bundles

03_FA_Uncinate_fasciculus_right: Fractional Anisotropy Uncinate fasciculus right. Units: NA. Coding: Fiber bundles

03_FA_Genu_of_corpus_callosum_right: Fractional Anisotropy Genu of corpus callosum right. Units: NA. Coding: Fiber bundles

03_FA_Body_of_corpus_callosum_right: Fractional Anisotropy Body of corpus callosum right. Units: NA. Coding: Fiber bundles

03_FA_Splenium_of_corpus_callosum_right: Fractional Anisotropy Splenium of corpus callosum right. Units: NA. Coding: Fiber bundles

03_MD_Corticospinal_tract_left: Mean Diffusivity Corticospinal tract left. Units: NA. Coding: Fiber bundles

03_MD_Anterior_limb_of_internal_capsule_left: Mean Diffusivity Anterior limb of internal capsule left. Units: NA. Coding: Fiber bundles

03_MD_Posterior_limb_of_internal_capsule_left: Mean Diffusivity Posterior limb of internal capsule left. Units: NA. Coding: Fiber bundles

03_MD_Posterior_thalamic_radiation_left: Mean Diffusivity Posterior thalamic radiation left. Units: NA. Coding: Fiber bundles

03_MD_Anterior_corona_radiata_left: Mean Diffusivity Anterior corona radiata left. Units: NA. Coding: Fiber bundles

03_MD_Superior_corona_radiata_left: Mean Diffusivity Superior corona radiata left. Units: NA. Coding: Fiber bundles

03_MD_Posterior_corona_radiata_left: Mean Diffusivity Posterior corona radiata left. Units: NA. Coding: Fiber bundles

03_MD_Cingulum_cingulate_gyrus_left: Mean Diffusivity Cingulum cingulate gyrus left. Units: NA. Coding: Fiber bundles

03_MD_Cingulum_hippocampus_left: Mean Diffusivity Cingulum hippocampus left. Units: NA. Coding: Fiber bundles

03_MD_Superior_longitudinal_fasciculus_left: Mean Diffusivity Superior longitudinal fasciculus left. Units: NA. Coding: Fiber bundles

03_MD_Superior_fronto_occipital_fasciculus_left: Mean Diffusivity Superior fronto occipital fasciculus left. Units: NA. Coding: Fiber bundles

03_MD_Inferior_fronto_occipital_fasciculus_left: Mean Diffusivity Inferior fronto occipital fasciculus left. Units: NA. Coding: Fiber bundles

03_MD_Uncinate_fasciculus_left: Mean Diffusivity Uncinate fasciculus left. Units: NA. Coding: Fiber bundles

03_MD_Genu_of_corpus_callosum_left: Mean Diffusivity Genu of corpus callosum left. Units: NA. Coding: Fiber bundles

03_MD_Body_of_corpus_callosum_left: Mean Diffusivity Body of corpus callosum left. Units: NA. Coding: Fiber bundles

03_MD_Splenium_of_corpus_callosum_left: Mean Diffusivity Splenium of corpus callosum left. Units: NA. Coding: Fiber bundles

03_MD_Corticospinal_tract_right: Mean Diffusivity Corticospinal tract right. Units: NA. Coding: Fiber bundles

03_MD_Anterior_limb_of_internal_capsule_right: Mean Diffusivity Anterior limb of internal capsule right. Units: NA. Coding: Fiber bundles

03_MD_Posterior_limb_of_internal_capsule_right: Mean Diffusivity Posterior limb of internal capsule right. Units: NA. Coding: Fiber bundles

03_MD_Posterior_thalamic_radiation_include_optic_radiation_right: Mean Diffusivity Posterior thalamic radiation include optic radiation right. Units: NA. Coding: Fiber bundles

03_MD_Anterior_corona_radiata_right: Mean Diffusivity Anterior corona radiata right. Units: NA. Coding: Fiber bundles

03_MD_Superior_corona_radiata_right: Mean Diffusivity Superior corona radiata right. Units: NA. Coding: Fiber bundles

03_MD_Posterior_corona_radiata_right: Mean Diffusivity Posterior corona radiata right. Units: NA. Coding: Fiber bundles

03_MD_Cingulum_cingulate_gyrus_right: Mean Diffusivity Cingulum cingulate gyrus right. Units: NA. Coding: Fiber bundles

03_MD_Cingulum_hippocampus_right: Mean Diffusivity Cingulum hippocampus right. Units: NA. Coding: Fiber bundles

03_MD_Superior_longitudinal_fasciculus_right: Mean Diffusivity Superior longitudinal fasciculus right. Units: NA. Coding: Fiber bundles

03_MD_Superior_fronto_occipital_fasciculus_right: Mean Diffusivity Superior fronto occipital fasciculus right. Units: NA. Coding: Fiber bundles

03_MD_Inferior_fronto_occipital_fasciculus_right: Mean Diffusivity Inferior fronto occipital fasciculus right. Units: NA. Coding: Fiber bundles

03_MD_Uncinate_fasciculus_right: Mean Diffusivity Uncinate fasciculus right. Units: NA. Coding: Fiber bundles

03_MD_Genu_of_corpus_callosum_right: Mean Diffusivity Genu of corpus callosum right. Units: NA. Coding: Fiber bundles

03_MD_Body_of_corpus_callosum_right: Mean Diffusivity Body of corpus callosum right. Units: NA. Coding: Fiber bundles

03_MD_Splenium_of_corpus_callosum_right: Mean Diffusivity Splenium of corpus callosum right. Units: NA. Coding: Fiber bundles

03_RD_Corticospinal_tract_left: Radial Diffusivity Corticospinal tract left. Units: NA. Coding: Fiber bundles

03_RD_Anterior_limb_of_internal_capsule_left: Radial Diffusivity Anterior limb of internal capsule left. Units: NA. Coding: Fiber bundles

03_RD_Posterior_limb_of_internal_capsule_left: Radial Diffusivity Posterior limb of internal capsule left. Units: NA. Coding: Fiber bundles

03_RD_Posterior_thalamic_radiation_left: Radial Diffusivity Posterior thalamic radiation left. Units: NA. Coding: Fiber bundles

03_RD_Anterior_corona_radiata_left: Radial Diffusivity Anterior corona radiata left. Units: NA. Coding: Fiber bundles

03_RD_Superior_corona_radiata_left: Radial Diffusivity Superior corona radiata left. Units: NA. Coding: Fiber bundles

03_RD_Posterior_corona_radiata_left: Radial Diffusivity Posterior corona radiata left. Units: NA. Coding: Fiber bundles

03_RD_Cingulum_cingulate_gyrus_left: Radial Diffusivity Cingulum cingulate gyrus left. Units: NA. Coding: Fiber bundles

03_RD_Cingulum_hippocampus_left: Radial Diffusivity Cingulum hippocampus left. Units: NA. Coding: Fiber bundles

03_RD_Superior_longitudinal_fasciculus_left: Radial Diffusivity Superior longitudinal fasciculus left. Units: NA. Coding: Fiber bundles

03_RD_Superior_fronto_occipital_fasciculus_left: Radial Diffusivity Superior fronto occipital fasciculus left. Units: NA. Coding: Fiber bundles

03_RD_Inferior_fronto_occipital_fasciculus_left: Radial Diffusivity Inferior fronto occipital fasciculus left. Units: NA. Coding: Fiber bundles

03_RD_Uncinate_fasciculus_left: Radial Diffusivity Uncinate fasciculus left. Units: NA. Coding: Fiber bundles

03_RD_Genu_of_corpus_callosum_left: Radial Diffusivity Genu of corpus callosum left. Units: NA. Coding: Fiber bundles

03_RD_Body_of_corpus_callosum_left: Radial Diffusivity Body of corpus callosum left. Units: NA. Coding: Fiber bundles

03_RD_Splenium_of_corpus_callosum_left: Radial Diffusivity Splenium of corpus callosum left. Units: NA. Coding: Fiber bundles

03_RD_Corticospinal_tract_right: Radial Diffusivity Corticospinal tract right. Units: NA. Coding: Fiber bundles

03_RD_Anterior_limb_of_internal_capsule_right: Radial Diffusivity Anterior limb of internal capsule right. Units: NA. Coding: Fiber bundles

03_RD_Posterior_limb_of_internal_capsule_right: Radial Diffusivity Posterior limb of internal capsule right. Units: NA. Coding: Fiber bundles

03_RD_Posterior_thalamic_radiation_include_optic_radiation_right: Radial Diffusivity Posterior thalamic radiation include optic radiation right. Units: NA. Coding: Fiber bundles

03_RD_Anterior_corona_radiata_right: Radial Diffusivity Anterior corona radiata right. Units: NA. Coding: Fiber bundles

03_RD_Superior_corona_radiata_right: Radial Diffusivity Superior corona radiata right. Units: NA. Coding: Fiber bundles

03_RD_Posterior_corona_radiata_right: Radial Diffusivity Posterior corona radiata right. Units: NA. Coding: Fiber bundles

03_RD_Cingulum_cingulate_gyrus_right: Radial Diffusivity Cingulum cingulate gyrus right. Units: NA. Coding: Fiber bundles

03_RD_Cingulum_hippocampus_right: Radial Diffusivity Cingulum hippocampus right. Units: NA. Coding: Fiber bundles

03_RD_Superior_longitudinal_fasciculus_right: Radial Diffusivity Superior longitudinal fasciculus right. Units: NA. Coding: Fiber bundles

03_RD_Superior_fronto_occipital_fasciculus_right: Radial Diffusivity Superior fronto occipital fasciculus right. Units: NA. Coding: Fiber bundles

03_RD_Inferior_fronto_occipital_fasciculus_right: Radial Diffusivity Inferior fronto occipital fasciculus right. Units: NA. Coding: Fiber bundles

03_RD_Uncinate_fasciculus_right: Radial Diffusivity Uncinate fasciculus right. Units: NA. Coding: Fiber bundles

03_RD_Genu_of_corpus_callosum_right: Radial Diffusivity Genu of corpus callosum right. Units: NA. Coding: Fiber bundles

03_RD_Body_of_corpus_callosum_right: Radial Diffusivity Body of corpus callosum right. Units: NA. Coding: Fiber bundles

03_RD_Splenium_of_corpus_callosum_right: Radial Diffusivity Splenium of corpus callosum right. Units: NA. Coding: Fiber bundles

04_Cerebrum_tcv: Total Cranial Volume* (Centimeters Cubed) *Cerebrum is the supratentorial portion of the cranium.. Units: NA. Coding: NA

04_Cerebrum_tcb: Total Brain Volume (Centimeters Cubed). Units: NA. Coding: NA

04_Cerebrum_tcc: Total Cerebrospinal Fluid Volume (Centimeters Cubed). Units: NA. Coding: NA

04_Cerebrum_gray: Total Gray Matter Volume (Centimeters Cubed). Units: NA. Coding: NA

04_Cerebrum_white: Total White Matter Volume (Centimeters Cubed). Units: NA. Coding: NA

04_Left_hippo: Segmented Left Hippocampus Volume (Centimeters Cubed). Units: NA. Coding: NA

04_Right_hippo: Segmented Right Hippocampus Volume (Centimeters Cubed). Units: NA. Coding: NA

04_Total_hippo: Segmented Total Hippocampi Volume (Centimeters Cubed). Units: NA. Coding: NA

04_Left_lateralvent: Segmented Left Lateral Ventricle Volume (Centimeters Cubed). Units: NA. Coding: NA

04_Right_lateralvent: Segmented Right Lateral Ventricle Volume (Centimeters Cubed). Units: NA. Coding: NA

04_Lateralvent: Segmented Total Lateral Ventricles Volume (Centimeters Cubed). Units: NA. Coding: NA

04_Thirdvent: Segmented Third Ventricle Volume (Centimeters Cubed). Units: NA. Coding: NA

04_Frontal_Cortical: Segmented Frontal Cortical Gray Matter Volume (Centimeters Cubed). Units: NA. Coding: NA

04_Occipital_Cortical: Segmented Occipital Cortical Gray Matter Volume (Centimeters Cubed). Units: NA. Coding: NA

04_Parietal_Cortical: Segmented Parietal Cortical Gray Matter Volume (Centimeters Cubed). Units: NA. Coding: NA

04_Temporal_Cortical: Segmented Temporal Cortical Gray Matter Volume (Centimeters Cubed). Units: NA. Coding: NA

04_Total_csf: Total Cerebrospinal Fluid Volume (Centimeters Cubed). Units: NA. Coding: NA

04_Total_gray: Total Gray Matter Volume (Centimeters Cubed). Units: NA. Coding: NA

04_Total_white: Total White Matter Volume (Centimeters Cubed). Units: NA. Coding: NA

04_Total_wmh: Total Cerebrum White Matter Hyperintensity Volume (Centimeters Cubed). Units: NA. Coding: NA

04_Total_brain: Total Brain Volume** (Centimeters Cubed) **This includes both cerebrum and infratentorial space (i.e., cerebellum and brainstem).. Units: NA. Coding: NA

01_AB_FBB_Present: AB_FBB Value present. Units: NA. Coding: {0:No}{1.00: Yes}

01_TAU_PI2620_Present: TAU_PI2620 Value present. Units: NA. Coding: {0:No}{1.00: Yes}

01_Hippocampal_Volume_Present: Hippocampal_Volume Value present. Units: NA. Coding: {0:No}{1.00: Yes}

01_Whole_Brain_Volume_Present: Whole_Brain_Volume Value present. Units: NA. Coding: {0:No}{1.00: Yes}

01_HHR_Present: HHR Value present. Units: NA. Coding: {0:No}{1.00: Yes}

01_metaROI_Present: metaROI Value present. Units: NA. Coding: {0:No}{1.00: Yes}

01_CorticalThickness_Present: CorticalThickness Value present. Units: NA. Coding: {0:No}{1.00: Yes}

01_MRI_Date: MRI Date. Units: NA. Coding: MM-YYYY

01_AB_FBB_Date: PET Amyloid Date. Units: NA. Coding: MM-YYYY

01_TAU_PI2620_Date: PET Tau Date. Units: NA. Coding: MM-YYYY

Examples

if (FALSE) { # \dontrun{
browseVignettes('HABSHD')
} # }