HD_subjinfo

HABS-HD dataset.

Usage

data(HD_subjinfo)

Format

A data frame with 3229 rows and 53 variables.

Details

Visit_ID: Visit Id number. Units: NA. Coding: Numeric

Med_ID: Medical Id number. Note: This is not an MRN and is not identifiable.. Units: NA. Coding: Numeric

Age: date of interview- date of birth. Units: (-9999 = N/A), (-8888 =Refused). Coding: Numeric

Ethnicity: Participant ethnicity. Units: (-9999 = N/A), (-8888 =Refused). Coding: Text

Interview_Language: Language in which the protocol was administered. Units: (-9999 = N/A), (-8888 =Refused). Coding: {1.00, English}{2.00, Spanish}

ID_Gender: Gender. Units: (-9999 = N/A), (-8888 =Refused). Coding: {.00, male}{1.00, female}

ID_Race_White: What is person's race? White?. Units: (-9999 = N/A), (-8888 =Refused). Coding: {.00, No}{1.00, yes}

ID_Race_Black: What is person's race? Black, African American, or Negro?. Units: (-9999 = N/A), (-8888 =Refused). Coding: {.00, No}{1.00, yes}

ID_Race_IndianAlaska: What is person's race? American Indian or Alaska Native? (specify below). Units: (-9999 = N/A), (-8888 =Refused). Coding: {.00, No}{1.00, yes}

ID_Race_Specify_AmerInd_Tribe: What is person's race? Asian, Indian, Chinese, Filipino, other Asian? (specify below). Units: (-9999 = N/A), (-8888 =Refused). Coding: {.00, No}{1.00, yes}

ID_Race_Asian: What is person's race? Japanese?. Units: (-9999 = N/A), (-8888 =Refused). Coding: {.00, No}{1.00, yes}

ID_Race_Specify_Asian_Race: What is person's race? Korean?. Units: (-9999 = N/A), (-8888 =Refused). Coding: {.00, No}{1.00, yes}

ID_Race_Japanese: What is person's race? Vietnamese?. Units: (-9999 = N/A), (-8888 =Refused). Coding: {.00, No}{1.00, yes}

ID_Race_Korean: What is person's race? Native Hawaiian?. Units: (-9999 = N/A), (-8888 =Refused). Coding: {.00, No}{1.00, yes}

ID_Race_Vietnamese: What is person's race? Guamanian or Chamorro?. Units: (-9999 = N/A), (-8888 =Refused). Coding: {.00, No}{1.00, yes}

ID_Race_NativeHawaiian: What is person's race? Samoan?. Units: (-9999 = N/A), (-8888 =Refused). Coding: {.00, No}{1.00, yes}

ID_Race_GuamChamorro: What is person's race? Other Pacific Islander? (specify below). Units: (-9999 = N/A), (-8888 =Refused). Coding: {.00, No}{1.00, yes}

ID_Race_Samoan: What is person's race? Other? (specify below). Units: (-9999 = N/A), (-8888 =Refused). Coding: {.00, No}{1.00, yes}

ID_Race_OtherPacific: Specify race. Units: (-9999 = N/A), (-8888 =Refused). Coding: Text

ID_Race_Specify_OtherPacific_Race: Specify American Indian tribe. Units: (-9999 = N/A), (-8888 =Refused). Coding: Text

ID_Race_Other: Specify Asian. Units: (-9999 = N/A), (-8888 =Refused). Coding: Text

ID_Race_Specify: Specify other pacific. Units: (-9999 = N/A), (-8888 =Refused). Coding: Text

ID_Hispanic: Is the person of Hispanic, Latino, or Spanish origin?. Units: (-9999 = N/A), (-8888 =Refused). Coding: {1.00, No, not hispanic}{2.00, Yes, Mexican American}{3.00, Yes, Puerto Rico}{4.00, Yes, Cuban}{5.00, Yes, other}

ID_Language_Primary: What is your primary language?. Units: (-9999 = N/A), (-8888 =Refused). Coding: {1.00, English}{2.00, Spanish}{3.00, Other}

MMSE_Total: MMSE Total score. Units: (-9999 = N/A), (-8888 =Refused). Coding: Numeric

SEVLT_DR_Total: SEVLT Trial Delayed Recall total score. Units: (-9999 = N/A), (-8888 =Refused). Coding: Numeric

LM2_AB_Total: WMS III Logical Memory 1 Story A+B Z Score. Units: (-9999 = N/A), (-8888 =Refused). Coding: Numeric

Digit_Symbol_Substitution: Digit Symbol Substitution total. Units: (-9999 = N/A), (-8888 =Refused). Coding: Numeric

Animal_Total: Animal Naming Total Score. Units: (-9999 = N/A), (-8888 =Refused). Coding: Numeric

CDX_Cog: Consensus Diagnosis - Cognitive Disorder. Units: (-9999 = N/A), (-8888 =Refused). Coding: {0, Cognitively Unimpaired}{1.00, Mild cognitive impairment}{2.00, Dementia}

ADI_StateRank: NA. Units: -9999. Coding: Group 1 is the lowest ADI and group 10 is the highest ADI. A block group with a ranking of 1 indicates the lowest level of "disadvantage" within the nation and an ADI with a ranking of 10 indicates the highest level of "disadvantage"

01_L_hippocampus: Whole hippocampal volume determined on a T1-weighted volume scan - left; control for intracranial volume (ICV_HD). Units: mm^3. Coding: Hippocampal_volume_scan_Release5 (1/2/3)

01_R_hippocampus: Whole hippocampal volume determined on a T1-weighted volume scan - right; control for intracranial volume (ICV_HD). Units: mm^3. Coding: Hippocampal_volume_scan_Release5 (1/2/3)

01_ICV_HD: Hippodeep- estimated intracranial volume determined on a T1-weighted volume scan. Please note that we previously used ICV from FreeSurfer. These two variables are slightly different and should not be mixed.. Units: mm^3. Coding: Hippocampal_volume_scan_Release5 (1/2/3)

01_AB_FBB_AB_pos: Amyloid positivity as defined by ADNI3 for FBB using a whole cerebellum reference region; Global SUVR >1.08. Units: NA. Coding: FBB_include_scan_Release5 (1/2)

01_TAU_PI2620_Medial_Temporal_SUVR: Medial Temporal SUVR - Mormino et al 2020 regions for 18F-PI-2620 (entorhinal cortex, hippocampus, amygdala); inferior cerebellar grey matter reference region. Units: NA. Coding: NA

02_WMH_Volume_Raw: White Matter Hyperintensity Volume. Units: NA. Coding: BorzageKing_R5_Combined

02_WMH_Volume_Log: NA. Units: NA. Coding: NA

APOE4_rs429358: rs429358 alleles, Implicated SNP in non-Hispanic White AD (Kunkle et al., 2019); nearest gene is APOE. Units: #. Coding: SNP Genotypes

APOE4_rs7412: rs7412 alleles. Units: #. Coding: SNP Genotypes

APOE4_SNP: APOE SNP GENTYPE. Units: #. Coding: SNP Genotypes

APOE4_Positivity: {0, Absent}{1, Present). Units: absent/present. Coding: SNP Genotypes

Examples

if (FALSE) {
browseVignettes('HABSHD')
}