HD_subjinfo

HABS-HD dataset.

Usage

data(HD_subjinfo)

Format

A data frame with 3838 rows and 52 variables.

Details

Visit_ID: Visit number. Units: -9999. Coding: Numeric

Med_ID: Medical ID (same as clinical). Units: -9999. Coding: Numeric

Age: Date of interview- Date of birth. Units: -9999. Coding: Numeric

Ethnicity: Participant ethnicity. Units: -9999. Coding: Text

ID_Gender: Gender. Units: -9999. Coding: {.00, male}{1.00, female}

Visit_ID: Visit Id number. Units: NA. Coding: Numeric

Med_ID: Medical Id number. Note: This is not an MRN and is not identifiable.. Units: NA. Coding: Numeric

Age: date of interview- date of birth. Units: (-9999 = N/A), (-8888 =Refused). Coding: Numeric

Interview_Language: Language in which the protocol was administered. Units: (-9999 = N/A), (-8888 =Refused). Coding: {1.00, English}{2.00, Spanish}

Ethnicity: Participant ethnicity. Units: (-9999 = N/A), (-8888 =Refused). Coding: Text

ID_Gender: Gender. Units: (-9999 = N/A), (-8888 =Refused). Coding: {.00, male}{1.00, female}

ID_Race_White: What is person's race? White?. Units: (-9999 = N/A), (-8888 =Refused). Coding: {.00, No}{1.00, yes}

ID_Race_Black: What is person's race? Black, African American, or Negro?. Units: (-9999 = N/A), (-8888 =Refused). Coding: {.00, No}{1.00, yes}

ID_Race_IndianAlaska: What is person's race? American Indian or Alaska Native? (specify below). Units: (-9999 = N/A), (-8888 =Refused). Coding: {.00, No}{1.00, yes}

ID_Race_Specify_AmerInd_Tribe: Specify American Indian tribe. Units: (-9999 = N/A), (-8888 =Refused). Coding: {.00, No}{1.00, yes}

ID_Race_Asian: What is person's race? Asian, Indian, Chinese, Filipino, other Asian? (specify below). Units: (-9999 = N/A), (-8888 =Refused). Coding: {.00, No}{1.00, yes}

ID_Race_Specify_Asian_Race: Specify Asian. Units: (-9999 = N/A), (-8888 =Refused). Coding: {.00, No}{1.00, yes}

ID_Race_Japanese: What is person's race? Japanese?. Units: (-9999 = N/A), (-8888 =Refused). Coding: {.00, No}{1.00, yes}

ID_Race_Korean: What is person's race? Korean?. Units: (-9999 = N/A), (-8888 =Refused). Coding: {.00, No}{1.00, yes}

ID_Race_Vietnamese: What is person's race? Vietnamese?. Units: (-9999 = N/A), (-8888 =Refused). Coding: {.00, No}{1.00, yes}

ID_Race_NativeHawaiian: What is person's race? Native Hawaiian?. Units: (-9999 = N/A), (-8888 =Refused). Coding: {.00, No}{1.00, yes}

ID_Race_GuamChamorro: What is person's race? Guamanian or Chamorro?. Units: (-9999 = N/A), (-8888 =Refused). Coding: {.00, No}{1.00, yes}

ID_Race_Samoan: What is person's race? Samoan?. Units: (-9999 = N/A), (-8888 =Refused). Coding: {.00, No}{1.00, yes}

ID_Race_OtherPacific: What is person's race? Other Pacific Islander? (specify below). Units: (-9999 = N/A), (-8888 =Refused). Coding: Text

ID_Race_Specify_OtherPacific_Race: Specify other pacific. Units: (-9999 = N/A), (-8888 =Refused). Coding: Text

ID_Race_Other: What is person's race? Other? (specify below). Units: (-9999 = N/A), (-8888 =Refused). Coding: Text

ID_Race_Specify: Specify other race. Units: (-9999 = N/A), (-8888 =Refused). Coding: Text

ID_Hispanic: Is the person of Hispanic, Latino, or Spanish origin?. Units: (-9999 = N/A), (-8888 =Refused). Coding: {1.00, No, not hispanic}{2.00, Yes, Mexican American}{3.00, Yes, Puerto Rico}{4.00, Yes, Cuban}{5.00, Yes, other}

ID_Language_Primary: What is your primary language?. Units: (-9999 = N/A), (-8888 =Refused). Coding: {1.00, English}{2.00, Spanish}{3.00, Other}

MMSE_Total: MMSE Total score. Units: (-9999 = N/A), (-8888 =Refused). Coding: Numeric

Digit_Symbol_Substitution: Digit Symbol Substitution. Units: (-9999 = N/A), (-8888 =Refused). Coding: Numeric

LM2_AB_Total: WMS lll Logical Memory 2 Story AB: Total. Units: (-9999 = N/A), (-8888 =Refused). Coding: Numeric

Animal_Total: Animal Name Scale Score. Units: (-9999 = N/A), (-8888 =Refused). Coding: Numeric

SEVLT_DR_Total: SEVLT Trial Delayed Recall Total Score. Units: (-9999 = N/A), (-8888 =Refused). Coding: Numeric

CDX_Cog: Cognitive Diagnosis. Units: (-9999 = N/A), (-8888 =Refused). Coding: {0, Cognitively Unimpaired}{1.00, Mild cognitive impairment}{2.00, Dementia}{9.00, Undetermined}

ADI_StateRank: Group 1 is the lowest ADI and group 10 is the highest ADI. A block group with a ranking of 1 indicates the lowest level of "disadvantage" within the nation and an ADI with a ranking of 10 indicates the highest level of "disadvantage". Units: (-9999 = N/A), (-8888 =Refused). Coding: Numeric

Visit_ID: Visit Number. Units: -9999. Coding: Numeric

Med_ID: Medical ID (same as clinical data). Units: -9999. Coding: Numeric

Age: Date of interview- Date of birth. Units: -9999. Coding: Numeric

Ethnicity: Participant ethnicity. Units: -9999. Coding: Text

ID_Gender: Gender. Units: -9999. Coding: {.00, male}{1.00, female}

APOE4_rs429358: rs429358 alleles, Implicated SNP in non-Hispanic White AD (Kunkle et al., 2019); nearest gene is APOE. Units: -9999. Coding: SNP Genotypes

APOE4_rs7412: rs7412 alleles. Units: -9999. Coding: SNP Genotypes

APOE4_Genotype: APOE SNP GENOTYPE. Units: -9999. Coding: SNP Genotypes

APOE4_Positivity: {0, Absent}{1, Present). Units: -9999. Coding: SNP Genotypes -{0: Absent}{1: Present)

Visit_ID: Visit number. Units: NA. Coding: Numeric

Med_ID: Medical ID (same as clinical). Units: NA. Coding: Numeric

Age: Date of interview- Date of birth. Units: NA. Coding: Numeric

Ethnicity: Participant ethnicity. Units: NA. Coding: Text

ID_Gender: Gender. Units: NA. Coding: {.00, male}{1.00, female}

01_L_hippocampus: Whole hippocampal volume determined on a T1-weighted volume scan - left; control for intracranial volume (ICV_HD). Units: NA. Coding: NA

01_R_hippocampus: Whole hippocampal volume determined on a T1-weighted volume scan - right; control for intracranial volume (ICV_HD). Units: NA. Coding: NA

01_ICV_HD: Hippodeep- estimated intracranial volume determined on a T1-weighted volume scan. Please note that we previously used ICV from FreeSurfer. These two variables are slightly different and should not be mixed.. Units: NA. Coding: NA

01_AB_FBB_AB_pos: Amyloid positivity as defined by ADNI3 for FBB using a whole cerebellum reference region; Global SUVR >1.08; PET scanner should be considered in statistical analyses.. Units: NA. Coding: NA

01_TAU_PI2620_Medial_Temporal_SUVR: Medial Temporal SUVR - Mormino et al 2020 regions for 18F-PI-2620 (entorhinal cortex, hippocampus, amygdala); inferior cerebellar grey matter reference region. Units: NA. Coding: NA

02_WMH_Volume_Raw: White Matter Hyperintensity Volume. Units: NA. Coding: NA

02_WMH_Volume_Log: NA. Units: NA. Coding: NA

Examples

if (FALSE) { # \dontrun{
browseVignettes('HABSHD')
} # }